[#60]/fix toasts util로 추출
